requirements.txt
python-telegram-bot
pandas
matplotlib
openai
sqlalchemy